First draft
This is just a small part of the first attempt of only using data in memory rather than editing database tables.

There are several features missing for this to be viable. This includes things like handling waypoints, disables and so on. There's also currently no way of removing spells from a trainer, only one to add a spell. This will hopefully be addressed in the future. Everything related to items also needs to be added.