Updates
Added the ability to Remove and update creature spawns.
Added the ability to alter creature templates.
Added the ability to Remove and update game object spawns. Still need to add a function to alter templates.
Added a function to remove vendor entries from NPCs. Still need to add the function to add entries.
Added the ability to change values of game events, making it possible to disable or alter events.