Fixes
Changed the creature template to only load the latest patch of a template since they do not need to be removed at any point.
Added a huge chunk of data for creature templates and spawns taken by comparing AzerothCore and vMaNGOS tables. It's not entirely accurate but is used as an example.