Fixes
Removed the counter for changes. It's really not necessary and gets confused if there are entries where values doesn't change.
Split the ifs for values into separate ifs instead of a joint one. There's no point in changing values that aren't wrong.
Added a first draft of adding items to vendors. I haven't really tested it yet but will do when I get to that point.
Removed the slot column from the vendor table as it's not used in the code anyway.