Rework the system
Began a rewrite to use hooks for the table queries instead of this approach. The current code will be saved on a separate branch, but will probably not be used in the future as it requires a lot of changes to the core.

I will also be using build numbers instead of a digit per patch so users may add their own patches between those provided here.

A huge thanks to kaytotes for helping out with this project!